+++ src/components/MenuSelect/MenuSelect.jsx
@@ -1,5 +1,5 @@
 import PropTypes from 'prop-types';
-import * as M from '../styles/MenuSelect.style';
+import * as M from '../MenuSelect/MenuSelect.style';
 
 function MenuSelect({ menus, activeMenu, setActiveMenu }) {
   return (
